refactor(room): 도메인 모델 추가
