+++ src/apps/rooms/models/requests.py
@@ -1,9 +1,9 @@
-from .responses import RoomItemResponse
+from .domain import RoomItem
 
 
-class RoomItemCreateRequest(RoomItemResponse):
+class RoomItemCreateRequest(RoomItem):
     """ 부동산 매물 정보 생성 요청"""
 
 
-class RoomItemUpdateRequest(RoomItemResponse):
+class RoomItemUpdateRequest(RoomItem):
     """ 부동산 매물 정보 수정 요청"""
